+++ src/main/java/cart/controller/view/HomeController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.view;
 
 import cart.service.ProductService;
 import lombok.RequiredArgsConstructor;
@@ -8,12 +8,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @RequiredArgsConstructor
 @Controller
-public class AdminPageController {
+public class HomeController {
     private final ProductService productService;
 
-    @GetMapping("/admin")
-    public String adminPage(Model model) {
+    @GetMapping("/")
+    public String indexPage(Model model) {
         model.addAttribute("products", productService.findAllProducts());
-        return "admin";
+        return "index";
     }
 }
