+++ src/test/java/cart/controller/HomeControllerTest.java
@@ -6,6 +6,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
+import cart.controller.view.HomeController;
 import cart.dto.ProductDto;
 import cart.entity.ProductEntity;
 import cart.service.ProductService;
